Separate Stream Renderer Plugin for the Render Delegate.
As part of separating the RendererPlugin from the RenderDelegate functionality, this change moves the Stream Render Delegate from the HdStream plugin into the HdSt library.  Replacing it with the RendererPlugin.

Even though the RenderDelegate has moved, code is still currently depending on the plugin discovery mechanism to pickup the RenderDelegate, so this means temporarily adding plugInfo.json to stream until we move all clients over to using the non-default constructor version of render index.

Changes:

- Moved HdStreamRenderDelegate to HdSt.

- Temporary added plugInfo.json to HdSt.
- Added HdStreamRendererPlugin, that creates a render delegate, but does not yet create a task controller.

(Internal change: 1727142)
(Internal change: 1727496)