Move tests that depend on Render Delegate behavior from Hd to HdSt.
Several tests in Hd depend on Render Delegate behavior, either because they generate an image or they test state that relies on prim syncing to have occurred.

Currently these test rely on the Render Delegate discovery mechanism in render index, which is going to be removed, so these tests need to be moved in order to remove that functionality.

This change is mostly mechanical operation of moving the files and updating the library name prefixes.

Changes:

- Moved Render Delegate based tests from Hd to HdSt.
- Renamed files and folders to update library name prefix.
- Updated build system files.

(Internal change: 1727567)
(Internal change: 1727762)