Connectability of shading attributes
Added connectability as token valued metadata on attributes.
In certain shading models, there is a need to distinguish between the
inputs can vary over a surface and those that must be uniform. This is
accomplished in USD-shade by limiting the connectability of the input, by
setting the "connectability" metadata on the associated attribute.

Connectability of an Input can be set to "full" or "interfaceOnly".
 * "full "implies that  the Input can be connected to any other Input or Output.
 * "interfaceOnly" implies that the Input can only be connected to a NodeGraph
 Input (which represents an interface override, not a render-time dataflow
 connection), or another Input whose connectability is also "interfaceOnly".

When connectability isn't authored on a shading node, it defaults to "full"
on a Shader and to "interfaceOnly" on a NodeGraph. When the prim type
is unknown, the fallback value (which is set in the schema registry) is "full".

Added UsdShadeConnectableAPI::CanConnect to help clients determine if a
input or output *can* be connected to a given shading attribute.
UsdShadeConnectableAPI::ConnectToSource() no longer validates the
the attempted connection. It simply goes ahead and tries to author it.
Clients are responsible for calling CanConnect() themselves when they
care about validation.

Unit test update is coming soon.

(Internal change: 1727857)