Fix crash when imaging gets an invalid texture
The problem is that in error conditions when reading a texture we
push back a null image.  However, on the client side, we are testing
for an error condition by checking that no images have been pushed back.

Also, turned some TF_CODING_ERRORs into TF_WARNs because they weren't
really an error in the code, but rather a failure to find a file, so
an asset-side issue.  Ideally these can be bubbled up to the application
as validation errors.

(Internal change: 1728372)