Refactor code to use new RenderIndex::New() entry point.
The current code base uses the default constructor on RenderIndex and the render delegate is either set later or it is discovered using the plugin mechanism, going forward code will need to provide the render delegate to the RenderIndex trough the RenderIndex::New() entry point.

This change ports all users of a RenderIndex in the tree over to the new entry point.  This has some side knock-on consequences, such as the Scene Delegate base no longer creates and manages a render index.  So the default constructor has been removed.

There was also an inconsistent within Hydra as if the RenderIndex was a raw pointers/reference or a shared_ptr.  As Hydra core no longer manages lifetimes of a render index, it was decided to use raw pointers in the API.  Allowing the user to use whatever method they wish too (Several tests use unqiue_ptr).

The general code pattern added right is:
- Create a RenderDelegate (typically HdStRenderDelegate directly).
- Construct a RenderIndex, passing the Render Delegate
- Construct a Scene Delegate, passing the Render Index.

In order to support tests in Hd, where no render delegate is defined a new "Null Render Delegate" has been created, that just implements the Render Delegate api, without providing any functionality.

Changes:

- Many code paths were using the default scene delegate constructor to create the render index.  These now implement the  construction pattern from above.
- Changing to this pattern meant a lot of Scene Delegates became heap rather than stack variables, so a lot of changing delegate. into delegate->
- Added unitTestNullRenderDelegate to Hd, used by HdTests.
- HdTestHelper uses the NullRenderDelegate, so it has been made private to prevent accidental use.
- Because of cross-module memory allocation concerns, HdRenderIndex::New has been made inline.
- Removed SceneDelegate's default constructor (marked as deleted along with the copy constructor and assignment operators)

- Removed UsdImagingDelegate default constructor (marked as deleted along with the copy constructor and assignment operators)
- Removed UnitTestDelegate default constructor
- ChangePointsAndTopoTest subtest of testHdRenderIndex disabled as this subtest needs to be moved to HdSt.
- Updated naming conventions in HdStUnitTestHelper after pre-commit branch.
- Converted over some Hdx test's to HdEngine's Execute method.
- UsdImagingGLHdEngine is now responsible for creating a Render Index for Usd.
- Removed old API version macros from Maya plugin.

(Internal change: 1728634)