Support lookup of file owner on Windows.
Previously we assumed the 'pwd' module was available but that's
not the case on Windows.  We now check platform.system() and on
Windows we use the 'win32security' module (from pywin32) if it's
available and otherwise the 'pwd' module.  On either platform,
we return '<unknown>' on error.

We're now also more careful about formatting the rollover text.
It uses HTML markup so we now escape '<', '>', and '&' in the
filename, username, and asset info values.

Fixes #118

(Internal change: 1729545)