Fix performance issues due to instance proxies
Use the prim path field in UsdObject for instance
proxies only. This field had been used for all objects
to simplify the code and avoid having multiple places
to check for the prim path. However, tests showed
that keeping this path updated made traversing prims
on a stage 2-4x slower, even when instance proxies
weren't involved.

The prim path in UsdObject is now only populated for
instance proxies and is left empty otherwise. Prim
traversal code was updated to accommodate this, but
much of this change is just renaming the field to
indicate that it's specific to proxies.

This ensures that consumers don't pay additional
performance costs unless they actually use the
instance proxy feature.

(Internal change: 1729775)