Fixed testArchStackTrace on Windows.
There doesn't seem to be a good way to catch heap corruption on
Windows (at least not one that can bypass the JIT debugger) so
this change simply disables those tests.

Also, we can't get symbolic stack traces in a release build so
rather than check in "main" is in the stack we pass if the back
trace isn't empty.

(Internal change: 1732258)