Fixed testArchTiming.
Previously this test would call ArchGetTickTime() and perform
a number of math operations to verify expected behavior.
However, if the result was large enough then integer overflow
would cause the verifications to fail.  In addition, any
failure is unreproducible.

Now the test tests a fixed arbitrary range of tick counts
starting at zero and tests each separately.  It also calls
ArchGetTickTime() twice separated by a sleep and verifies
that the number of ticks approximately corresponds to the
sleep time.

This indicated that the code to find the clock frequency on
Windows was broken, off by about a factor of 1000.  This code
was replaced with code that counts the number of ticks in some
measured interval.

(Internal change: 1732260)