Fix for possible buffer overrun while calculating GPU smooth normals.
Fix a buffer overrun which was caused by the smooth normal shader to read off the end of the adjacency table, leading to random valence values, which are used for both a loop count, with the inner loop doing additional indirect indexing into arrays causing additional buffer over- and under-runs.

These large loops and buffer over-/under-runs was stalling the system in what appears to be PCI buslock behavior.  Causing not just the GPU, but the CPU to stall out as well.

The cause of the initial overrun was how many times we run the smooth normal shader.  The current code run the shader once for every vertex (point) in the input buffer.  However, there is an optimization when calculating the adjacency table, that looks at the mesh topology and calculates the maximum vertex index in the index buffer and uses that, thus avoiding the need to compute adjacency/normals for vertices not used by the mesh.

The CPU smooth normal computation, took this into account and clamped the invocations, but the GPU computation did not.

Changes:

- Clamp the number of runs of the GPU smooth normal shader to the size of the adjacency table.

(Internal change: 1732303)