[Hdx, Hd, UsdImagingGL] Expose per-element picking via UsdImagingGL API
The first part adds elementIds to the old and the new imaging path,
this allows usdview to potentially pick elements when HD_ENABLED=0 or 1.

The second part refactors GfVec4i and uint8_t[] based decodes of
id buffer values from color to use a single implementation called
DecodeIDRenderColor. Hopefully this is a first step towards refactoring
the id render pipeline completely to make non-GL implementations
easier in the future.

Lastly a new API is added to fetch Rprim paths from the render index
using a decoded id. This is used internally in imaging and usdImaging.

(Internal change: 1733763)