Stop using 't' in fopen mode strings.
fopen() mode strings should not contain 't'.  That's an extension
and it's implied;  'b' is used to indicate a binary file.  There
were two uses of 't' in tf that intended a text file so the 't'
was dropped.

There was one use of 't' in sdf that was replaced by 'b'.
textFileFormat.yy gets the file length then reads that many bytes.
But on Windows the length will be 2 for each '\r\n' while fread()
counts each as 1 because '\r\n' is replaced with '\n'. This caused
an unexpected end-of-file error.

Fixes #189.

(Internal change: 1734230)