Fixed GarchNSGLContextState.
There was an assignment that was supposed to be an equality
comparison.  This caused a TF_VERIFY error on exit.  It also
messed up the internal state on contexts, which would have
been a problem if we had multiple contexts that hashed to
the same value.

This change also removed an unused member variable and changed
a C-style cast to C++-style.

(Internal change: 1735076)