Use _stat64() on Windows to handle large files.
All uses of stat() in C++ should go into arch but for now this
handles uses outside arch and tf and those in tf/fileUtils.cpp. The
calls to stat() in glf and ar were replaced by calls to arch
functions, including a new version of ArchGetModificationTime() that
takes a path instead of a struct stat.

One use in fileUtils.cpp detects if a file exists.  This is called by
the resolver to see if a resolved path exists but on Windows this
would fail for files over 4GB (2GB?) because the size wouldn't fit in
the struct stat.  Windows supports large files through _stat64() and
struct __stat64 so now we use that.

(Internal change: 1735378)