Update doxygen config to ignore namespace macros
doxygen was combining the PXR_NAMESPACE_OPEN_SCOPE
macro with the next doxygen-tagged symbol it saw.
If that symbol were a function, it would consider
it part of the signature and include it in that
function's documentation. If it were another macro
it would fail to generate documentation with:

warning: documentation for unknown define
PXR_NAMESPACE_OPEN_SCOPEAR_DECLARE_RESOLVER_CONTEXT found.

Per the doxygen FAQ, this is due to the fact that we
typically don't close uses of the namespace macros
with semi-colons. This is fixed by telling doxygen's
preprocessor to remove these macros from consideration.

(Internal change: 1736193)