Improve memory utilization in adjacency table.
We have a large prim in a show asset that contains a very large vertex valence value, which was causing the size of the adjacency table to exceed the maximum buffer size supported by Hydra.

Previously the size of the entries in the adjacency table was based of the maximum vertex valence, to give us a fixed entry pitch.  While simpler, this utilized a lot more memory.

This change re-organizes the adjacency table, so the first part of the table contains a pair of entries for each index.  The first being the offset into the table when the vertex, next and previous pairs are stored.  The second value being the vertex valence.

Modifying the table this way, has the nice property of changing the table's alignment from 4-bytes to 8-bytes.

So this change effective adds 4-bytes to the table for every entry.  So, theoretically can increase memory use, if every vertex has the same valence value.  However, in practice this change has been showing massive memory savings.  For example, In the shot for which the bug was originally reported to total size of all adjacency tables was reduced from 8GB to 1.5GB.

This change does introduce an extra level of indirection when looking up adjacency values.  So there is some performance concerns.  However, as the buffer is more tightly packed, it is hopped that improved cache performance would offset the cost of the additional indirection.  Performance testing has not relieved any significant slowdowns.

Changes:
- Updated the adjacency table builder computation, CPU smooth normal computation and GPU smooth normal shader for the new table format.
- The size of the adjacency table, no longer reflex the number of points in the table, so this information is tracked along with the table.
- Added the size of the adjacency table as a performance metric.
- Removed adjacencyStride and padding from Gpu Smooth Normal's Uniforms.

(Internal change: 1736970)