Keep context for WorkRunDetachedTask() alive forever.
It's unsafe to use a static object for the WorkRunDetachedTask()
context because it can be accessed after main exits.  At that
point it might already be destroyed.  Or it might be created for
the first time then soon destroyed by the C++ static local d'tor
handler.  Either way we create a task using a bogus or soon-to-be
bogus context.

This caused an intermittent failure in testSdfAttributeBlocking_Cpp.
~SdfData calls WorkSwapDestroyAsync() and this happens after or
just before exiting main.  The WorkRunDetachedTask context is soon
destroyed, often before the task completes.  When that happens TBB
raises an assertion.

So now we create the context on the heap and never clean it up.

(Internal change: 1737947)