Stopped using concurrent_wait in UsdGeomBBoxCache.
Using concurrent_wait means wait_for_all() will leave the root
task's ref count at 1 so we don't have to set it back to 1 after
the wait.  But it also means any child task that uses wait_for_all()
will also end with a ref count of 1, something that we don't expect.
This isn't a problem in WorkDispatcher because we don't expose the
root or children tasks so clients can't create grandchildren tasks.

However, a client can create a new root task in the context of a
WorkDispatcher task.  If it does that without supplying its own
task_group_context then the root task (and of its descendant tasks)
will inherit the WorkDispatcher's context, including the
concurrent_wait trait.  That means they'll unexpectedly have a ref
count of 1 after wait_for_all().  If a task doesn't handle this by
the time it finishes then the parent will assertion failure on the
non-zero ref count (in a debug TBB build).

This was happening in UsdGeomBBoxCache.  So now we create the new
root task with a new context without concurrent_wait.

(Internal change: 1738037)