Don't author targets to external flattened masters
When processing relationship targets in an instance,
flattening would author targets to the special "flattened
master" prims, even though the instance referenced that
prim. This was inconsistent with how the targets were
represented in the original, unflattened scene. Now,
flattening will author targets that are relative to
the instance.

(Internal change: 1738316)