Clarify that a deliberately missing symbol is okay.
On Windows the TestPlugDsoUnloadable test has a symbol that is
deliberately missing.  Unfortunately the linker will report this
as an error even though we tell it to link anyway.  This error
can be confusing to users.

This change tries to explain the situation more clearly and it
changes the name of the missing symbol to something the user
might read and realize that the error can be ignored.

Fixes #199.

(Internal change: 1738391)