Don't translate targets and connections to masters
UsdRelationship and UsdAttribute no longer translate
authored target/connection paths that point to objects
beneath instances to the corresponding object in the
instances' masters. This is no longer needed since
consumers can use instance proxies to get Usd
objects for these paths; previously, consumers were
only able to work with the masters.

The 'forwardToObjectsInMasters' argument has been
removed, which simplifies the API. This removes an
odd case where using the target/connection API on
a relationship/attribute in a master and passing
false for this argument would return results based
on whatever prim index was selected as the master's
source, which was useful only in very limited cases
and was more likely to be misused by consumers.

(Internal change: 1738477)