Remove error for targets/connections in instances pointing to instance root
Previously, Usd considered a relationship target or attribute
connection inside an instance that pointed to the instance root to
be invalid. UsdRelationship and UsdAttribute would issue errors
if it detected this case, but only if the Usd object being used was
beneath a prim in a master. This error would not be issued if the
Usd object was beneath an instance proxy.

This is clearly inconsistent, and we don't believe this construct
is fundamentally invalid. So, we're removing the error.

(Internal change: 1738511)