Enhanced browsing of instances in usdview Prim browser.
Now that we have core support for instance proxy prims, it
is trivial to expand instances inline in the browser.  Interestingly,
making instance proxies selectable automatically makes displayed
selected BBOX's "just work" for sub-instance geometry, even though
usdImaging hiliting does not.  Made a small change to
UsdImagingDelegate::PopulateSelection() so that, given a "proxy path",
it will now hilite the entire instance to which the proxy belongs,
which is much better than hiliting nothing, which is what would
otherwise have happened.

At the same time, we'd like to make it a little less easy to enable
"Show Masters", because:
a) it will now rarely provide any additional useful information
b) it can be a very expensive option to enable, because masters
   are all top-level prims (so all get processed imediately),
   and the number of masters will affect the sensitivity of the
   browser's scroll bar.

To that end, we remove the "Jump to Master" command and menu
entries, because its action enables "Show Masters" for all.  In the
future, we will provide more useful, selection-based commands like
"Select all instances like me" (i.e. all instances that share the
same master), etc.

Also, to try to purge "Show Masters" from peoples' state files,
we renamed the action to actionShow_All_Master_Prims so it
gets a fresh entry.

(Internal change: 1738566)