[Tf] Update stopwatch test to only check lower bounds.
std::sleep_for only gaurantees that it sleeps for at least
some amount of time. It gives no promises as to the upper bound.
Given this, we should only be testing on the lower bound.
This test became a problem on a vm which has much longer sleep times
on average.

(Internal change: 1738928)