Fix composition bug with multiple relocations
Pcp was not composing opinions for a relocated prim
if the site where the relocation was authored was also
relocated in the same layer stack. This was because
the relocation map Pcp used to determine the source
of the relocated prim baked the multiple relocations
that affected it into a single entry, throwing out the
intermediate sites in the layer stack that could also
have opinions for that prim.

To fix this, composition now uses an "incremental"
relocation map stored in PcpLayerStack that does
not bake entries together.

(Internal change: 1742335)