[Core, Hd] H2P: lighting/normal differences between Glim and Newimg
Prevented Glim from quadrangulating a triangle only surface when in LowCompexity.
Also fixed the normal interpolation in the case where complexity > Low
The edge veritces where not normalized and the center normal was not rescaled.
Results now closer match OSD results.

Added Ptex tests to ensure we can verify what parts of forced quadrangulation we
cover and to get informed when coverage changes.

The escond part of the bug reported needs AlphaToSampleCoverage to work in
Presto with new imaging.

More Glim tests were also investigated and are in a second CL. We can invest more
time in it if needed.

Added a Presto based Ptex test as well that can be used with Glim or Hydra.

(Internal change: 1743411)