Fixed ArchLibrary*() calls and, by extension, TfDl*() calls.
ArchLibraryClose() should return 0 on success and non-zero on
error;  on Windows it was the opposite.

ArchLibraryError() should return the error from the most recent
call to ArchLibraryOpen() or ArchLibraryClose() or zero if the last
such call was successful.  Previously on Windows it was getting the
last error from any Win32 API call and not returning 0 after
successful calls.

Finally, the TfDl test constructed a path to a test library with a
"lib" prefix.  That's only used on Unix-like systems.  Arch now
defines an ARCH_LIBRARY_PREFIX which is "" on Windows and "lib" on
Unix-like systems and TfDl now uses this macro.

(Internal change: 1743835)