Fixed TfPyNoticeWrapper.
TfPyNoticeWrapper inherited a TfNotice type and TfPyNoticeWrapperBase
in that order.  But the registered Tf_PyObjectFinderBase for the
type would static_cast a void* to the most derived type to a
TfPyNoticeWrapperBase*.  That void* points, of course, to the
TfNotice in TfPyNoticeWrapper, not the TfPyNoticeWrapperBase.
We now cast to TfPyNoticeWrapper, which was probably the original
intent.

(Internal change: 1743849)