Fixed testTfPython on Windows.
There were a few problems.  First, string(nullptr) simply crashes
on Windows rather than raising the expected exception so now we
directly raise an exception.  Second, on Windows sys.__stdout__
does not necessarily use the same FILE* as stdout though writes
to both go to the same underlying OS file handle.  The same is
true of sys.__stderr__.  We now do translation of these FILE
pointers in the SetOutputFile() wrapper since that function will
check that the FILE* is stdout or stderr.  Finally, testTfPython
itself was misusing tempfile.mkstemp().  That function returns an
open file handle and its path but the test was not closing that
handle before trying to remove the file by its path.  That fails
on Windows.  Now we use NamedTemporaryFile() which automatically
deletes the file on close and allows the use of a context manager
to close the file cleanly.

(Internal change: 1743850)