Fixed race condition in TfType.
TfTypeInfoMap::Find() was potentially modifing its data members.
Therefore multithreaded clients must use an exclusive mutex when
calling it but TfType did not -- it only held a shared reader lock.
Given the name of the method and that it was const, it's no
surprise this mistake was made.

That version of Find() was removed.  This will cause a performance
hit if it has to fall back to string-based lookup frequently but it
won't crash.  There's also a new version of Find() that takes a
functor that ensures the thread has exclusive access to the object.
This is used by TfType and could be used by any client.

This change also removes 'mutable' from the data members and makes
methods const/non-const as appropriate.

(Internal change: 1743852)