Fixed testTfScriptModuleLoader on Windows.
The problem here was stdout buffering.  This change does two things.
1) TfDebug now flushes its stream after every write.  This makes
sense since we if we're asking for debug output we probably don't
want to lose it in case of a crash.  2) testTfScriptModuleLoader.py
now makes stdout unbuffered.  On Unix this doesn't matter but on
Windows it appears that the Python and C++ buffers for stdout are
different when output is being redirected instead of going to a tty.
As a result you'd see a chunk of C++ output then a chunk of Python
output and so on.  This caused the diff to fail.

(Internal change: 1743853)