Catch Incorrectly sized Constant Data.
A test asset was exported incorrectly, such that it had a constant color interpolation mode, but was providing an array of data to Hydra.

This is condition was caught late in Hydra, when the data was transferred to the GPU.  At which point much information about the source of the data had been lost.

This change, adds an earlier check to make sure that if the primvar is marked as constant, it only has a single element and if not a warning is reported and the data source is not added to the prim.

Changes:
- In HdRprim::_PopulateConstantPrimVars(), issued a TF_WARN if the sources num elements is not 1.

(Internal change: 1743911)