Fixes for documentation on Windows.
We need quotes around @DOT_EXECUTABLE@ in case the path has
spaces it (which is almost certain on Windows and possible on
Unix).

Using TemporaryFile() as an output appears to be problematic
on Windows, possibly due to exclusive file locking.  This
changes the doxygen invokation to collect output via a pipe.

(Internal change: 1744307)