Fixed ArchDemangle() replacement of std::string typename.
ArchDemangle() has code to replace the RTTI name for std::string
with 'string' to make the demangled type name more readable.  But
this code was searching for and replacing the mangled named.  On
Linux this is 'Ss', on Windows 'class std::basic_string<char,
struct std::char_traits<char>,class std::allocator<char> >' and on
Mac 'NSt3__112basic_stringIcNS_11char_traitsIcEENS_9allocatorIcEEEE'.

So on Windows this works correctly because the RTTI name is already
demangled, on Linux it screws up any demangled type name with 'Ss'
in it, and on Mac it does nothing at all because it never matches.

Now we get and demangle the RTTI name for std::string and compare
against that.

(Internal change: 1745909)