CMake cleanup in preparation for static library support.
This doesn't change how shared libraries for tests are built,
nor does it change whether we're building libraries as shared
or static.  The only difference in the build results should be
that we get fewer paths in each rpath.

Unified cmake code to build shared, static and plugin libraries.
Eliminated some unnecessary functions.

Fixed two libraries putting plugInfo.json in with PYTHON_FILES
instead of RESOURCE_FILES.

The Maya plugin now uses a new option, MAYA_PLUGIN, to indicate
that it's a Maya plugin.  Previously the plugin function had the
name of plugin hardcoded.

_plugInfo_subst() no longer expects PLUG_INFO_LIBRARY_PATH to be
set and instead requires it as an argument.

Added functions for building and setting rpaths and reduced the
rpath to the minimum.

(Internal change: 1747246)
(Internal change: 1747475)