Fix relocates regressions from change 1742335
Map functions for relocates were being created with
multiple entries for the same target prim. This made
the map functions non-invertible and caused bad path
translations when processing inherits, which resulted
in spurious composition errors. PcpLayerStack now
checks for this case and will favor 'combined' relocation
entries, which should provide all the info needed for
path translations.

In addition, nodes for incremental relocations were
introduced, which led to more implied inherit nodes.
This was expected; however, this led to relocates
being evaluated for placeholder implied inherit
nodes, which was not expected and caused failed
verifies. This work should not be necessary, so the
fix here is simply to avoid evalulating relocates
in this case.

(Internal change: 1747503)