[usdMaya] add assembly representation namespace handling based on Autodesk sample code
After discussion with Autodesk, we've decided to adopt the namespace
handling functionality from their sample assembly reference
implementation here:

http://help.autodesk.com/view/MAYAUL/2017/ENU/?guid=__cpp_ref_scene_assembly_2assembly_reference_8h_example_html

This should really be implemented internally as built-in functionality
of Maya assemblies rather than having to deal with it in plugin code,
but there are currently no plans to make that happen, so we're forced to
do it ourselves. This helps ensure that assembly edits do not fall off
when assembly nodes are duplicated/renamed/etc.

(Internal change: 1747905)