Next step in supporting static libraries.
There is now a BUILD_SHARED_LIBS option that defaults to ON.
When OFF we build individual static libraries instead of
dynamic libraries.  Since this prevents Python modules and
plugins from working those aren't built in that case.  In
addition any test written in Python is dropped as are any
tests marked with the new REQUIRES_PYTHON_MODULES option.
The mayapy tests use a new CUSTOM_PYTHON argument when
being registered.  The value is the path to the Python to
use to run the test and is also used to detect that the test
uses Python.

Linking against static libraries means pulling in the entire
archive, not just symbols that are used, because global
constructors including ARCH_CONSTRUCTOR functions can have
effects.  So we now use _pxr_target_link_libraries() instead
of target_link_libraries() everywhere except when building a
library.  This allow us to modify link lines appropriately to
pull in all symbols.  _pxr_target_link_libraries()

Because we want to choose shared or static based on the
BUILD_SHARED_LIBS option, it no longer makes sense to add a
library using pxr_shared_library().  There's now pxr_library()
that adds a shared or static library based on the option.
pxr_shared_library() remains and adds a shared library even
when building static but there are no callers.  The static
counterpart, pxr_static_library() also remains and is used
by the Kanata libraries.

This change also makes a few other tweaks to ensure we're
picking up the right libraries.  Finally, two new macros are
defined: PXR_PYTHON_ENABLED is always defined for now but can
be not defined if/when we support not linking Python at all;
PXR_PYTHON_MODULES_ENABLED is defined only if we're building
Python modules so it's not defined when BUILD_SHARED_LIBS is
OFF.  This allows tests (currently just testVtCpp) to skip
portions of the test that require Python modules.

Finally, this change has the beginnings of support for a
monolithic library where all libraries are linking into one
library but it doesn't support that yet.

(Internal change: 1748436)