Don't remove duplicates when determining internal and external libraries.
This fixes an error when _pxr_target_link_libraries
is called with a list of libraries with configurations
embedded in it, e.g. "debug;lib_d.so;optimized;lib.so".
If the same library were specified for different
configurations, the de-duplication would wind up
passing something like "debug;lib.so;optimized" to
target_link_libraries, which would result in an error.

(Internal change: 1748801)