Fix for not initializing Repr's on Prims under certain collection use pattern.
My previous change for dirty bit propagation revealed a bug with the handling of repr initialization.

The problem was caused by the following usage pattern:
- Render with Dirty List 1 and Repr1.
- Insert Prim
- Render with Dirty List 2 and Repr2.
- Render with Dirty List 1 and Repr1 again.

The issue was the new prim wasn't synced with Repr1.
The current code handles this by doing a sync when gathering Draw Items, however this violates scene
delegate usage policy (pulling data that's not marked as dirty).

This change fixes this use case and another potential use case issue where the repr was modified on a collection.

Changes:
- The change tracker already tracks a version number for when the rprim set in the render index changes,
  this change exposes that value.
- The Dirty List now tracks the version of the render index and the repr of the collection it last synced with.
- If the render index version change or the collection's repr changed, we do a forced sync of the prims
  so that they get a change to initialize with the new repr.

(Internal change: 1750524)