Pre-Sync Dirty Bit Propagation.
This is a recommit of the "Pre-Sync Dirty Bit Propagation" change now that the repr initialization bug is fixed.  This change incorporates some feedback from a post-commit review, where the code can fetch the type of the points from the points resource rather than storing it separately.

Validity of the points resource was an issue and to handle that case, when points are detected to be invalid, the prim's bar is removed and the primvar sync aborted.  Several follow up bugs have been filed to investigate issues around primvar metadata changes that arose from this change.

Changes from Original Commit:

- If the points primvar is invalid, reset the prim's bar to empty and early exit.

- Removed points valid and type members from HdStMesh.
- Gpu smooth normals will first attempt to get the type of points from any new points primvar buffer source.  If there is not a new points primvar,
it will then attempt to get the type from a points resource in the bar.  If that fails, the gpu smooth normal computation isn't added.

Original Commit Comment:

This issue started with a report that the debug build of TBB was throwing errors because of an invalid iterator when deleting a value

from usdImaging's value cache.  After some digging this turned out to be a double delete of the points primvar.

To get a double delete, it must mean that we were requesting the same primvar twice from the cache, which breaks the scene delegate

policy of only pulling data once.  The cause of this double-pull was down to the points data being invalid.  However, the GPU smooth normal path had code in it that would pull the points primvar if it wasn't found when looping over all the dirty primvars.

This in itself is troublesome, even in the non-error case, because this meant that we pulling the points primvar even if the data wasn't dirty.  Which breaks another scene delegate policy of only pulling data which was marked as dirty in an earlier call to the scene delegate.

While investigating this bug, a second bug report came in, which was a crash in the smooth normals GPU computation execution.  This crash was because the points buffer was null.  Again, the problem was caused by invalid points data, but we were trying to run the smooth normals computation anyway.  This change addresses this issue as well.

Going deeper into the rabbit hole, I also discover another case where we pull data from the scene delegate without it being dirty in the sync

request vector and that is when we encounter a new repr for a Rprim.

The primary change in this fix to add a new pre-sync stage to render index sync.  This stage visits all dirty Rprims and performs two operations that

can cause additional dirty bits to be set for an Rprim.  The first operation is to Initialize the Repr's for the Rprim.  The second is to Propagate dirty

bits (i.e. certain dirty bits cause others to be set).  As each Rprim is independent in this regard, this operation is performed in parallel.

This occurs before the Sync Request vector is passed to the scene delegates, (So the scene delegate has the correct set of dirty bits).

Previously, as _GetRepr() would initialize new Repr's and thus detect when a new Repr was created.  This was passed down to other sync functions via the "isNew" parameters.  To handle this case, A new custom dirty bit was added, DirtyNewRepr, which is set when _InitRepr() creates a new repr.  Later functions can then use this dirty bit rather than the parameter.

Changes:

- Added _PreSyncRPrims worker thread function, which calls InitRepr and PropagateRprimDirtyBits on the Rprim.

- Added _PreSyncRequestVector worker thread function, which calls _PreSyncRPrims on each rprim (Parrellel for) for a specific scene delegate.

- Added Pre-Sync Rprims to RenderIndex's SyncAll() method, with dispatches to _PreSyncRequestVector for each scene delegate.

- Rprim's _GetRepr function can fail, so added error handling/propagation up the stack (to avoid null ptr referencing).

- For Each Rprim type the Repr initalization was factored out of _GetRepr() into _InitRepr().

- Each Rprim's _InitRepr() sets the DirtyNewRepr custom dirty bit when creating a new repr.  _GetRepr() clears it.

- HdStMesh's _PopulateVertexPrimVars, no longer takes isNew param and removed the unused "desc" param.

- HdStMesh's _PopulateVertexPrimVars(), no longer has special code to process points primvar even when it's not dirty.

- HdStMesh, now tracks the valid state of the points primvar.  Smooth Normal's are no longer computed if the points data is invalid.

- The reason for the second pull of the points primvar in GPU smooth normals was to get the type of points, so now this information is

  remembered by the class.

- CPU smooth normal computation requires the points primvar, so this is now handled by dirty bit propagation to ensure the points primvar

  is marked dirty before requesting it from the delegate.

- Refactored some of the smooth normal code to get rid of duplicate code between cpu and gpu paths.

(Internal change: 1751291)