Fix usdImaging not honoring its contract on imageable prim instances, which sometimes causes bad artifacts.
Before, when an imageable prim was marked as instanceable we tried to render it, this means we were not being complient with USD and we were also rendering incorrect results.

Now, when a imageable prim is instanced we output a warning and we do not render it, we do advice the user to create a transform and instance that transform as a solution.

(Internal change: 1754374)